# Replaced concat with push for performance