## Description
Replaces instances of `d = d.concat(...)` with `push` within encoding functions for improved performance per #512.  Should not impact results. 

## Motivation and Context
This version is faster, particularly on Firefox.  As a test, I ran `window.font.download()` in the console of the https://opentype.js.org/ site.  In Firefox the change reduced runtime from 1,341ms to 190ms.  Chrome experienced a more modest improvement from 203ms to 166ms.  The ability to quickly generate custom font files is important for my project [Scribe OCR](https://github.com/Balearica/scribeocr), and I'm sure others as well. 

## How Has This Been Tested?
The unit tests pass and I am using this version in my live site.  Additionally, the change is conceptually simple and narrow in scope. 

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->


## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] I did `npm run test` and all tests passed green (including code styling checks).
- [ ] I have added tests to cover my changes.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the **README** accordingly.
- [x] I have read the **CONTRIBUTING** document.