@yne 2022-05-21
It seems related to the fact that
`.concat()` create an intermediate array
`.push()` mutate the existing array

Very nice improvement :+1: